requirements.txt
email-validator==2.2.0
flask-login==0.6.3
flask==3.1.1
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
flask-wtf==1.2.2
werkzeug==3.1.3
wtforms==3.2.1
openai==1.78.1
sqlalchemy==2.0.41
python-dotenv==1.1.0
python-magic==0.4.27
redis==5.0.0 